Started a table of contents
Up to now just the commands to puzzle a table
together from section, subsection and
card-entries.

The subsection still has a minor error, if a
document starts with a subsection.

Still missing is an automatic way to put the
together while defining the cards itself and a 
proper output after all cards.