Changed the drawing of grid lindes - again…
to fit the request, that if you resize the page on
print, the outer lines will also be visible, hence
each card now has a complete border. That made
some shift parameters depending on the format
necessary, but they were easy to be obtained just
by measurements in the PDF.

Despite that the borders are nearly \cardwidth
in width (\cardheight in height) because if they
would, the borders would still vanish.
It might also work to just subtract an pt, but
for now it's too late to change that, maybe i'll
try tomorrow.